chore: Restore openapi_extra for create endpoint
